July 26th AFP episode
This is the utilities project we began writing in the July 16th 2018 episode of the Atlanta
Functional Programming Common Lisp study group session. Here we defined a new ASDF system with a
package that uses defpackage-plus to re-export some useful functions from the "alexandria" and
"serapeum" libraries that we talked about. In the future, we will also define and export our own
utilities.